refactor: Consolidate payment modal logic into shared utilities and remove unused configurations
