--- frontend/micro-ui/web/micro-ui-internals/packages/modules/hearings/src/configs/MakeSubmissionConfig.js
@@ -1,1 +1,0 @@
-export { configs } from "@egovernments/digit-ui-module-home/src/configs/MakeSubmissionConfig";
