+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/orders/src/pages/employee/PaymentForSummonModal.js
@@ -1,4 +1,4 @@
-import React, { useMemo, useState, useCallback } from "react";
+import React, { useMemo, useState, useEffect } from "react";
 import { RadioButtons, CardLabel, LabelFieldPair } from "@egovernments/digit-ui-react-components";
 import { useTranslation } from "react-i18next";
 import CustomToast from "@egovernments/digit-ui-module-dristi/src/components/CustomToast";
@@ -10,29 +10,21 @@ import usePaymentProcess from "../../../../home/src/hooks/usePaymentProcess";
 import { DRISTIService } from "@egovernments/digit-ui-module-dristi/src/services";
 import { ordersService } from "../../hooks/services";
 import { Urls } from "../../hooks/services/Urls";
-import { useEffect } from "react";
 import { paymentType } from "../../utils/paymentType";
 import { DateUtils, extractFeeMedium, getAuthorizedUuid, getTaskType } from "@egovernments/digit-ui-module-dristi/src/Utils";
-import { getAdvocates } from "../../utils/caseUtils";
+import {
+  formatRespondentAddressLine,
+  getViewOrderClickHandler,
+  useCaseLockStatusForPaymentModal,
+  useIsUserAdvocateOnCase,
+} from "./shared/paymentSummonModalShared";
 import ButtonSelector from "@egovernments/digit-ui-module-dristi/src/components/ButtonSelector";
 import { getPartyNameForInfos } from "../../utils";
 import { CaseWorkflowState } from "@egovernments/digit-ui-module-dristi/src/Utils/caseWorkflow";
 import { ORDER_TYPES } from "../../utils/constants";
+import { submitModalInfoPostPayment as submitModalInfo } from "./shared/paymentSummonPostReceiptDefaults";
 
 const modeOptions = [{ label: "E-Post (3-5 days)", value: "e-post" }];
-
-const submitModalInfo = {
-  header: "CS_HEADER_FOR_SUMMON_POST",
-  subHeader: "CS_SUBHEADER_TEXT_FOR_Summon_POST",
-  caseInfo: [
-    {
-      key: "Case Number",
-      value: "FSM-2019-04-23-898898",
-    },
-  ],
-  isArrow: false,
-  showTable: true,
-};
 
 const PaymentForSummonComponent = ({
   infos,
@@ -162,36 +154,9 @@ const PaymentForSummonModal = ({ path }) => {
   }, [caseData]);
 
   const caseCourtId = useMemo(() => caseDetails?.courtId, [caseDetails]);
-  const fetchCaseLockStatus = useCallback(async () => {
-    try {
-      const status = await DRISTIService.getCaseLockStatus(
-        {},
-        {
-          uniqueId: caseDetails?.filingNumber,
-          tenantId: tenantId,
-        }
-      );
-      setIsCaseLocked(status?.Lock?.isLocked);
-    } catch (error) {
-      console.error("Error fetching case lock status", error);
-      const errorId = error?.response?.headers?.["x-correlation-id"] || error?.response?.headers?.["X-Correlation-Id"];
-      setShowToast({ label: t("ERROR_FETCHING_CASE_LOCK_STATUS"), error: true, errorId });
-    }
-  });
-  useEffect(() => {
-    if (caseDetails?.filingNumber) {
-      fetchCaseLockStatus();
-    }
-  }, [caseDetails?.filingNumber]);
+  useCaseLockStatusForPaymentModal(caseDetails, tenantId, t, setIsCaseLocked, setShowToast);
 
-  const allAdvocates = useMemo(() => getAdvocates(caseDetails), [caseDetails]);
-  const advocatesUuids = useMemo(() => {
-    if (allAdvocates && typeof allAdvocates === "object") {
-      return Object.values(allAdvocates).flat();
-    }
-    return [];
-  }, [allAdvocates]);
-  const isUserAdv = useMemo(() => advocatesUuids.includes(authorizedUuid), [advocatesUuids, authorizedUuid]);
+  const isUserAdv = useIsUserAdvocateOnCase(caseDetails, authorizedUuid);
 
   const todayDate = new Date().getTime();
   const dayInMillisecond = 24 * 3600 * 1000;
@@ -483,12 +448,7 @@ const PaymentForSummonModal = ({ path }) => {
 
   const infos = useMemo(() => {
     const addressDetails = filteredTasks?.[0]?.taskDetails?.respondentDetails?.address;
-    const formattedAddress =
-      typeof addressDetails === "object"
-        ? `${addressDetails?.locality || ""}, ${addressDetails?.city || ""}, ${addressDetails?.district || ""}, ${addressDetails?.state || ""}, ${
-            addressDetails?.pincode || ""
-          }`
-        : addressDetails;
+    const formattedAddress = formatRespondentAddressLine(addressDetails);
     return [
       { key: "Issued to", value: getPartyNameForInfos(orderDetails, compositeItem, orderType) },
       { key: "Next Hearing Date", value: DateUtils.getFormattedDate(new Date(hearingsData?.HearingList?.[0]?.startTime), "DD-MM-YYYY") },
@@ -504,12 +464,7 @@ const PaymentForSummonModal = ({ path }) => {
   }, [hearingsData?.HearingList]);
 
   const links = useMemo(() => {
-    const onViewOrderClick = () => {
-      history.push(
-        `/${window.contextPath}/citizen/dristi/home/view-case?caseId=${caseData?.criteria?.[0]?.responseList?.[0]?.id}&filingNumber=${filingNumber}&tab=Orders`
-      );
-    };
-    return [{ text: "View order", link: "", onClick: onViewOrderClick }];
+    return [{ text: "View order", link: "", onClick: getViewOrderClickHandler({ history, caseData, filingNumber }) }];
   }, [caseData?.criteria, filingNumber, history]);
 
   const paymentForSummonModalConfig = useMemo(() => {
